requirements.txt
objectpath==0.6.1
python-dotenv==0.19.1
requests==2.26.0
terminaltables==3.1.10